requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
langchain-community
google-genai
ragas
python-dotenv
pandas
torch
sentence-transformers
jsonref
ipykernel
ipywidgets